chore: add strict lint gate and clean workspace warnings
Add a top-level Makefile with discoverable test, lint, audit, and
release-check targets for production-readiness work.

Clean the existing workspace warnings so strict linting can be used as
a release/CI gate instead of accepting warning drift. This includes the
tx_wait unused assignment, test-only dead-code warnings, and related
clippy cleanup across workspace crates.

Tests: make test
Lint: make lint-strict
Build: cargo build --release -p sum-node

+++ crates/sum-node/src/upload.rs
@@ -409,7 +409,7 @@ impl UploadOrchestrator {
 
         let chunks_fully_confirmed = per_chunk_confirmations
             .iter()
-            .filter(|&&count| count >= REPLICATION_FACTOR as u32)
+            .filter(|&&count| count >= REPLICATION_FACTOR)
             .count() as u32;
 
         Ok(UploadResult {
@@ -568,7 +568,7 @@ mod tests {
     /// Every chunk fully replicated to R archives → check_success returns Ok.
     #[test]
     fn check_success_happy_path_full_replication() {
-        let r = REPLICATION_FACTOR as u32;
+        let r = REPLICATION_FACTOR;
         let result = make_result(vec![r, r, r, r, r], false, vec![], r);
         assert!(result.check_success(r).is_ok());
     }
@@ -577,7 +577,7 @@ mod tests {
     /// with that chunk's index reported.
     #[test]
     fn check_success_under_replicates_single_chunk() {
-        let r = REPLICATION_FACTOR as u32;
+        let r = REPLICATION_FACTOR;
         // Chunk 5 only got R-1 ACKs.
         let mut per_chunk = vec![r; 10];
         per_chunk[5] = r - 1;
@@ -602,7 +602,7 @@ mod tests {
     /// flag means the result is `Err(Timeout)`.
     #[test]
     fn check_success_timeout_dominates() {
-        let r = REPLICATION_FACTOR as u32;
+        let r = REPLICATION_FACTOR;
         let result = make_result(vec![r, r, 0, 0], true, vec![], r);
         assert!(matches!(result.check_success(r), Err(UploadFailure::Timeout)));
     }
@@ -611,7 +611,7 @@ mod tests {
     /// counts — the orchestrator already knows something went wrong.
     #[test]
     fn check_success_failed_push_is_fatal() {
-        let r = REPLICATION_FACTOR as u32;
+        let r = REPLICATION_FACTOR;
         let failed = vec![FailedPush {
             chunk_index: 7,
             cid: "bafk_test_chunk7".to_string(),
@@ -634,7 +634,7 @@ mod tests {
     /// claim success while leaving chunks under-replicated.
     #[test]
     fn check_success_recomputes_fully_confirmed_ignoring_cache() {
-        let r = REPLICATION_FACTOR as u32;
+        let r = REPLICATION_FACTOR;
         // Real per-chunk state: chunk 2 is missing one ACK.
         let mut per_chunk = vec![r; 4];
         per_chunk[2] = r - 1;
@@ -666,7 +666,7 @@ mod tests {
     /// ascending order.
     #[test]
     fn check_success_mixed_under_replication_indices() {
-        let r = REPLICATION_FACTOR as u32;
+        let r = REPLICATION_FACTOR;
         // Chunks 0, 3, 8 are under-replicated.
         let per_chunk = vec![
             r - 1, r, r, r - 2, r, r, r, r, 0, r,
